refector : 카드 이미지 상수데이터 분리
