+++ three session/test/src/card.jsx
@@ -1,0 +1,12 @@
+const cardImages = [
+    { src: "./img/가렌1.jpg", matched: false },
+    { src: "./img/다이애나1.jpg", matched: false },
+    { src: "./img/요네1.jpg", matched: false },
+    { src: "./img/유미1.jpg", matched: false },
+    { src: "./img/티모1.jpg", matched: false },
+    { src: "./img/그웬1.jpg", matched: false },
+    { src: "./img/케이틀린.jpg", matched: false }
+  ];
+  
+  export default cardImages;
+
